break the Steps from antd into components
